requirements.txt
# Echo Bot - Discord AI Worldbuilding Assistant
# Phase 1+ Dependencies - Multi-AI Arsenal

# Discord Integration
discord.py>=2.3.2

# Local AI (Ollama/Mistral)
ollama>=0.1.6

# Memory Layer
mem0ai>=0.0.9

# External AI APIs
anthropic>=0.18.0
google-generativeai>=0.4.0
openai>=1.12.0

# Configuration & Environment
python-dotenv>=1.0.0

# Utilities
aiohttp>=3.9.1
asyncio>=3.4.3

